Refactor script to accept cmd line args
Refactored the build script to add some command line options. I
wanted a "clean" option to be able to clean up everything without
having to actually bulid. Moved the clean and build operations
into functions.

The alpine-make-rootfs script is currently pointing to one that
is in my own repo and it disables root login. There is a pr
opoen to merge this into the main branch. Will move back to the
main alpine branch at some point.